requirements.txt
playwright
pandas
numpy
requests
python-dotenv